Implemented test case for neural network.

+++ src/valuation/models/linear_regression_torch_model.py
@@ -8,7 +8,9 @@ class LRTorchModel(nn.Module):
         super().__init__()
         self.n_input = n_input
         self.n_output = n_output
-        tt_A = torch.tensor(np.random.normal(0, 0.02, size=(n_output, n_input)))
+        tt_A = torch.tensor(
+            np.random.normal(0, 0.02, size=(n_output, n_input)), dtype=torch.float32
+        )
         self.A = nn.Parameter(tt_A, requires_grad=True)
 
     def forward(self, x: torch.tensor) -> torch.tensor:
